moved options to dedicated folder
