+++ sdk/storage/Azure.Storage.Blobs/perf/Azure.Storage.Blobs.Perf/Options/IBlobClientOptionsOptions.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Azure.Storage.Blobs.Perf.Infrastructure
+namespace Azure.Storage.Blobs.Perf.Options
 {
     public interface IBlobClientOptionsOptions
     {
